GPL-3 is better after all
See:
http://kbroman.org/pkg_primer/pages/licenses.html

If you choose the GPL, note that there are multiple versions.
GPLv2 is the old version. Don’t choose that. Choose the newer
one, GPLv3. The newer one fixes some loopholes in the older one.

To use the GPLv3 with your R package, include the following line
in your DESCRIPTION file.

License: GPL-3

That’s it.